Move BRAMQueue to targetutils
