+++ sim/midas/src/main/scala/midas/core/FPGAManagedStreamEngine.scala
@@ -9,6 +9,7 @@ import org.chipsalliance.cde.config.Parameters
 import freechips.rocketchip.diplomacy._
 
 import midas.widgets._
+import midas.targetutils.{BRAMQueue}
 
 class WriteMetadata(val numBeatsWidth: Int) extends Bundle {
   val numBeats = Output(UInt(numBeatsWidth.W))
